chore: update to utils function
